# backend3/docs/swagger.yaml
basePath: /api
definitions:
...
  oneops_backend3_model_cmdb.BusinessUnit:
    properties:
      children:
        description: 关联
        items:
          $ref: '#/definitions/oneops_backend3_model_cmdb.BusinessUnit'
        type: array
      code:
        description: 业务代码
        type: string
      createdAt:
        type: string
      email:
        description: 联系邮箱
        type: string
      id:
        type: integer
      level:
        description: 层级
        type: integer
      name:
        description: 业务名称
        type: string
      owner:
        description: 负责人
        type: string
      parentId:
        description: 父业务ID
        type: integer
      phone:
        description: 联系电话
        type: string
      remarks:
        description: 备注
        type: string
      sortOrder:
        description: 排序
        type: integer
      status:
        description: 状态：1启用 0禁用
        type: integer
      updatedAt:
        type: string
    type: object
  oneops_backend3_model_cmdb.Cabinet:
    properties:
      capacity:
        description: U数
        type: integer
      code:
        description: 机柜代码
        type: string
      createdAt:
        type: string
      id:
        type: integer
      name:
        description: 机柜名称
        type: string
      position:
        description: 位置
        type: string
      powerCapacity:
        description: 总电力(KW)
        type: number
      powerUsage:
        description: 已用电力(KW)
        type: number
      remarks:
        description: 备注
        type: string
      room:
        allOf:
        - $ref: '#/definitions/oneops_backend3_model_cmdb.ServerRoom'
        description: 关联
      roomId:
        description: 机房ID
        type: integer
      servers:
        items:
          $ref: '#/definitions/oneops_backend3_model_cmdb.Server'
        type: array
      status:
        description: 状态：1启用 0禁用
        type: integer
      updatedAt:
        type: string
      usedU:
        description: 已用U数
        type: integer
    type: object
  oneops_backend3_model_cmdb.CloudServer:
    properties:
      chargeType:
        description: postpay/prepay
        type: string
      createdAt:
        type: string
      id:
        type: integer
      instanceId:
        description: 云主机实例ID
        type: string
      instanceName:
        description: 实例名称
        type: string
      instanceType:
        description: 实例规格
        type: string
      privateIp:
        description: 内网IP
        type: string
      provider:
        description: aliyun, tencent, aws, huawei
        type: string
      publicIp:
        description: 公网IP
        type: string
      region:
        description: 地域
        type: string
      securityGroups:
        description: 安全组JSON
        type: string
      server:
        allOf:
        - $ref: '#/definitions/oneops_backend3_model_cmdb.Server'
        description: 关联
      serverId:
        type: integer
      subnetId:
        description: 子网ID
        type: string
      updatedAt:
        type: string
      vpcId:
        description: VPC ID
        type: string
      zone:
        description: 可用区
        type: string
    type: object
...
  oneops_backend3_model_cmdb.CredentialType:
    enum:
    - user
    - system
    type: string
    x-enum-comments:
      CredentialTypeSystem: 用于系统自动化运维（Agent 部署/采集）
      CredentialTypeUser: 用于用户堡垒连接
    x-enum-descriptions:
    - 用于用户堡垒连接
    - 用于系统自动化运维（Agent 部署/采集）
    x-enum-varnames:
    - CredentialTypeUser
    - CredentialTypeSystem
  oneops_backend3_model_cmdb.DiskPartition:
    properties:
      mount:
        description: 挂载点，如 /var, /home
        type: string
      usage:
        description: 使用率百分比
        type: number
    type: object
  oneops_backend3_model_cmdb.SSHCredential:
    properties:
      authType:
        $ref: '#/definitions/oneops_backend3_model_cmdb.SSHPublicKeyAuth'
      createdAt:
        type: string
      credentialType:
        allOf:
        - $ref: '#/definitions/oneops_backend3_model_cmdb.CredentialType'
        description: user | system
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      passphrase:
        description: 加密存储
        type: string
      password:
        description: 加密存储
        type: string
      port:
        type: integer
      privateKey:
        description: 加密存储
        type: string
      servers:
        description: 关联
        items:
          $ref: '#/definitions/oneops_backend3_model_cmdb.Server'
        type: array
      sortOrder:
        type: integer
      status:
        type: integer
      updatedAt:
        type: string
      username:
        type: string
    type: object
  oneops_backend3_model_cmdb.SSHPublicKeyAuth:
    enum:
    - password
    - key
    type: string
    x-enum-comments:
      SSHAuthKey: 密钥认证
      SSHAuthPassword: 密码认证
    x-enum-descriptions:
    - 密码认证
    - 密钥认证
    x-enum-varnames:
    - SSHAuthPassword
    - SSHAuthKey
  oneops_backend3_model_cmdb.Server:
    properties:
      agentPort:
        type: integer
      agentStatus:
        description: uninstalled | running | offline
        type: string
      agentVersion:
        type: string
      arch:
        description: 系统架构
        type: string
      assetNumber:
        description: 资产编号
        type: string
      attributes:
        items:
          $ref: '#/definitions/oneops_backend3_model_cmdb.ServerAttribute'
        type: array
      business:
        $ref: '#/definitions/oneops_backend3_model_cmdb.BusinessUnit'
      businessId:
        type: integer
      cabinet:
        $ref: '#/definitions/oneops_backend3_model_cmdb.Cabinet'
      cabinetId:
        description: 所在机柜ID
        type: integer
      cloudInfo:
        $ref: '#/definitions/oneops_backend3_model_cmdb.CloudServer'
      connectivityStatus:
        description: 连通性状态
        type: string
      cpu:
        description: CPU核心数
        type: integer
      cpuUsage:
        type: number
      createdAt:
        type: string
      credentialId:
        description: SSH凭证ID（兼容旧字段）
        type: integer
      credentialIds:
        items:
          type: integer
        type: array
      credentialNames:
        description: 凭证名称JSON数组
        type: string
      credentials:
        items:
          $ref: '#/definitions/oneops_backend3_model_cmdb.SSHCredential'
        type: array
      disk:
        description: 磁盘(GB)
        type: integer
      diskPartitions:
        items:
          $ref: '#/definitions/oneops_backend3_model_cmdb.DiskPartition'
        type: array
      diskUsage:
        type: number
      env:
        description: 环境
        type: string
      expireWarranty:
        description: 保修到期
        type: string
      groupIds:
        items:
          type: integer
        type: array
      groupNames:
        description: 冗余字段：优化列表查询性能（避免关联查询）
        type: string
      groups:
        items:
          $ref: '#/definitions/oneops_backend3_model_cmdb.ServerGroup'
        type: array
      hostname:
        description: 主机名
        type: string
      id:
        type: integer
      innerIp:
        description: 内网IP
        type: string
      instanceId:
        description: 云主机实例ID
        type: string
      instanceType:
        description: 云主机类型
        type: string
      ip:
        description: 外网IP
        type: string
      lastCheckTime:
        description: 最后连通性检查时间
        type: string
      lastConnectTime:
        description: 最后连接时间
        type: string
      lastHeartbeatAt:
        type: string
      load1:
        type: number
      load5:
        type: number
      load15:
        type: number
      manufacturer:
        description: 厂商
        type: string
      memory:
        description: 内存(GB)
        type: integer
      memoryUsage:
        type: number
      metricsUpdatedAt:
        type: string
      model:
        description: 型号
        type: string
      os:
        description: 操作系统
        type: string
      osVersion:
        description: 系统版本
        type: string
      provider:
        description: 服务商
        type: string
      purchaseDate:
        description: 购买日期
        type: string
      region:
        description: 区域
        type: string
      remarks:
        description: 备注
        type: string
      serverType:
        description: 类型
        type: string
      sn:
        description: 序列号
        type: string
      sshCredential:
        allOf:
        - $ref: '#/definitions/oneops_backend3_model_cmdb.SSHCredential'
        description: 关联
      sshCredentialId:
        description: SSH凭证ID
        type: integer
      sshPort:
        description: SSH端口
        type: integer
      status:
        description: 状态
        type: string
      systemCredential:
        $ref: '#/definitions/oneops_backend3_model_cmdb.SSHCredential'
      systemCredentialId:
        description: 系统运维凭证（Agent部署/采集专用）
        type: integer
      tags:
        items:
          $ref: '#/definitions/oneops_backend3_model_cmdb.ServerTag'
        type: array
      uPosition:
        description: 机柜位置(U)
        type: integer
      updatedAt:
        type: string
      zone:
        description: 可用区
        type: string
    type: object
  oneops_backend3_model_cmdb.ServerAttribute:
    properties:
      attributeId:
        type: integer
      attributeKey:
        type: string
      attributeValue:
        type: string
      category:
        type: string
      createdAt:
        type: string
      definition:
        $ref: '#/definitions/oneops_backend3_model_system.AttributeDefinition'
      id:
        type: integer
      server:
        allOf:
        - $ref: '#/definitions/oneops_backend3_model_cmdb.Server'
        description: 关联
      serverId:
        type: integer
      updatedAt:
        type: string
      valueType:
        type: string
    type: object
  oneops_backend3_model_cmdb.ServerGroup:
    properties:
      children:
        items:
          $ref: '#/definitions/oneops_backend3_model_cmdb.ServerGroup'
        type: array
      code:
        description: 分组代码
        type: string
      color:
        description: 分组颜色
        type: string
      createdAt:
        type: string
      description:
        description: 分组描述
        type: string
      icon:
        description: 分组图标
        type: string
      id:
        type: integer
      level:
        description: 层级
        type: integer
      name:
        description: 分组名称
        type: string
      parent:
        allOf:
        - $ref: '#/definitions/oneops_backend3_model_cmdb.ServerGroup'
        description: 关联
      parentId:
        description: 父分组ID
        type: integer
      servers:
        items:
          $ref: '#/definitions/oneops_backend3_model_cmdb.Server'
        type: array
      sortOrder:
        description: 排序
        type: integer
      status:
        description: 状态：1启用 0禁用
        type: integer
      updatedAt:
        type: string
    type: object
  oneops_backend3_model_cmdb.ServerRoom:
    properties:
      address:
        description: 详细地址
        type: string
      cabinets:
        description: 关联
        items:
          $ref: '#/definitions/oneops_backend3_model_cmdb.Cabinet'
        type: array
      code:
        description: 机房代码
        type: string
      contact:
        description: 联系人
        type: string
      createdAt:
        type: string
      id:
        type: integer
      location:
        description: 机房位置
        type: string
      name:
        description: 机房名称
        type: string
      phone:
        description: 联系电话
        type: string
      provider:
        description: 服务商
        type: string
      remarks:
        description: 备注
        type: string
      status:
        description: 状态：1启用 0禁用
        type: integer
      updatedAt:
        type: string
    type: object
  oneops_backend3_model_cmdb.ServerTag:
    properties:
      color:
        description: 标签颜色
        type: string
      createdAt:
        type: string
      description:
        description: 标签描述
        type: string
      id:
        type: integer
      name:
        description: 标签名称
        type: string
      servers:
        description: 关联
        items:
          $ref: '#/definitions/oneops_backend3_model_cmdb.Server'
        type: array
      sortOrder:
        description: 排序
        type: integer
      status:
        description: 状态：1启用 0禁用
        type: integer
    type: object
...
  oneops_backend3_model_system.AttributeDefinition:
    properties:
      category:
        type: string
      createdAt:
        type: string
      defaultValue:
        type: string
      description:
        type: string
      id:
        type: integer
      key:
        type: string
      name:
        type: string
      options:
        type: string
      required:
        type: boolean
      sortOrder:
        type: integer
      status:
        type: integer
      type:
        type: string
      updatedAt:
        type: string
    type: object
...
  oneops_backend3_pkg_utils.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8082
info:
  contact: {}
  description: |-
    OneOps 一体化运维平台后端 API
    包含系统管理、CMDB、K8s、审计、监控、应用授权等模块
  title: OneOps API
  version: "1.0"
paths:
...
  /system/attributes:
    get:
      description: 获取属性定义列表，支持按分类筛选
      parameters:
      - description: 属性分类
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取属性列表失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 获取属性定义列表
      tags:
      - CMDB-属性
    post:
      consumes:
      - application/json
      description: 创建一个新的属性定义（需要管理员权限）
      parameters:
      - description: 属性定义信息
        in: body
        name: attr
        required: true
        schema:
          $ref: '#/definitions/oneops_backend3_model_system.AttributeDefinition'
      produces:
      - application/json
      responses:
        "200":
          description: 请求参数错误 / 无权限 / 创建失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 创建属性定义
      tags:
      - CMDB-属性
  /system/attributes/{id}:
    delete:
      description: 根据属性 ID 删除属性定义（需要管理员权限）
      parameters:
      - description: 属性 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 无效的 ID / 无权限 / 删除失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 删除属性定义
      tags:
      - CMDB-属性
    get:
      description: 根据属性 ID 获取属性定义详情
      parameters:
      - description: 属性 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 无效的 ID / 属性不存在
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 获取属性定义详情
      tags:
      - CMDB-属性
    put:
      consumes:
      - application/json
      description: 根据属性 ID 更新属性定义信息（部分字段更新，需要管理员权限）
      parameters:
      - description: 属性 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 需要更新的字段
        in: body
        name: attr
        required: true
        schema:
          $ref: '#/definitions/oneops_backend3_model_system.AttributeDefinition'
      produces:
      - application/json
      responses:
        "200":
          description: 无效的 ID / 请求参数错误 / 无权限 / 更新失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 更新属性定义
      tags:
      - CMDB-属性
...
  /system/server-attributes/{serverId}:
    get:
      description: 根据主机 ID 获取主机的属性列表
      parameters:
      - description: 主机 ID
        in: path
        name: serverId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 无效的主机 ID / 获取主机属性失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 获取主机属性列表
      tags:
      - CMDB-属性
    post:
      consumes:
      - application/json
      description: 保存指定主机的属性列表（覆盖更新）
      parameters:
      - description: 主机 ID
        in: path
        name: serverId
        required: true
        type: integer
      - description: 属性列表
        in: body
        name: attributes
        required: true
        schema:
          items:
            $ref: '#/definitions/oneops_backend3_model_cmdb.ServerAttribute'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 无效的主机 ID / 请求参数错误 / 保存失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 保存主机属性
      tags:
      - CMDB-属性
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token}，token 由 /api/login 获取
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"